+++ CHANGELOG.md
@@ -8,6 +8,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - MV-190 - Show always opened tip after two hours
 - MV-89 - Do not take account of disabled type/category (zero result) in catalog filter
 - MV-182 - Show amount included taxes when event does not have taxes
+### Added
+- VIMEET-2019 - View multiple sheets contacts list; Impersonate as every user of a participation sheet
 
 ## [2.11.0] - 2019-08-27
 ### Added